Update to Polymer 0.2.0.
I duplicated rules involving '^' so they'd work in both the polyfill and native
Shadow DOM.

I used the 'attached' event instead of 'ready' to be sure elements are in the tree.

I delayed traversals outside of the current element with 'this.async()' because
apparently the currently world upgrades in a less strict order.

Drive-by fix to use this.super() for upcalls.

I fixed a template in titlepage.html where I'd forgotten a '}'.