Update func.html
Remove the redundant overloads of `operator==` that have been struck from C++20.  For further simplicity, this follows the convention fully enforced in C++23 that the functions declared in the header synopsis are not repeated below the class definition.

I believe this change to be editorial, as we explicitly reference C++20 for the specification here, so this is just aligning to match up the declarations, there are no normative wording changes.